refactor: share the per-app QRS reads between run and plan
getSheetsTaggedWith and the whole pre-session QRS block (app metadata
with its not-found contract, both tagged-sheet lookups, the repo/engine
sheet-id map) move into qseow-tagged-sheets.js as readQseowAppContext,
called by qseowProcessApp and the dry-run planner alike. The dry run's
promise is 'the same reads, in the same order' - a promise two copies
had already started to break within one PR.
